Rewrite database structure and indexing
The new database structure uses two related tables, containing files
and variables, respectively.  This significantly reduces duplication
and allows for more efficient querying.  To go alongside this, the
indexing has been rewritten to always use distributed workers and try
to run more efficiently: database operations are performed on the
workers rather than the broker; and we avoid set operations on the
full list of output files.

Instead of assuming a file structure like
<rootdir>/<configuration>/<experiment>/<run>, we denote the parent
directory of a given run as the `experiment.'  This should be
sufficiently unique, and allows flexibility in indexing arbitrary
directories.

To make querying for date ranges and interrogating the data coverage
easier, we add some file metadata about time information.

For querying, it's going to be significantly easier to abstract a
little bit over the database by using sqlalchemy.  Additionally, my
assumption that there wouldn't be much lock contention for database
writes was probably wrong... Instaed, we just use the futures
mechanism to return data back to the main thread for database
writes (we could open per-worker connections, but we'll probably hit
concurrency issues again...)