Allow mean of time bounds as alternative to time dimension
In some cases (e.g. CICE), time-averaged data is saved with the time
dimension at the end of the averaging period.  In the case of CICE,
this actually lies wholly outside the averaging period, and leads to a
slight offset: monthly data appears in the first day of the following
month.  By passing from_bounds to get_nc_variable(), the mean of the
time bounds used for averaging is used instead.  In our example, this
means that monthly averages appear in the middle of the
month (matching MOM5 behaviour).