Remove assumption on payu directory structure
This has a few follow-on changes, like making a distributed Client
optional for indexing. It may still be advantageous to perform
distributed indexing, but push the parallel map to the loop over
individual files in index_experiment(). This way, all the database
querying for updates can be done first, thus maintaining the
constraints for accessing a database session.