Make querying.getvar only return a DataArray
With the advances in xarray, we have ended up duplicating some of its
functionality. Paring down the data to a certain time range can be
done on the resulting DataArray, and rebasing the times is less
necessary due to more intuitive handling of CFTimes. If rebasing is
required, it's available separately.