Don't index empty output files with a time dimension
A NetCDF file with a record (time) dimension of length 0 is probably
an error. In particular, we don't know any time start/end or frequency
information in these files, and they don't concatenate properly,
causing errors later down the line.

Even though these files are valid, we now consider them as "not
present", so they're not included in queries, and will be picked up to
be re-indexed by the pruning mechanism.