Use a cascading search to determine the time dimension (#145)
* Use a cascading search to determine the time dimension

The assumption that time was the only record dimension in a file was
not particularly robust against real data. We instead check:
- standard_name == 'time'
- axis == 'T'
- record dimension
- name.lower() == 'time'

Fixes #144.

* Check dimension in dataset variables

A dimension without an associated variable causes a KeyError when
we're looking for dimensions that contain a particular attribute, when
this should just count as not finding the attribute.