Don't create new tables on outdated databases
We were telling the ORM to create all tables before the database
version check, which means that if create_session was run on an
outdated database, it would still get the new structure. We've added a
few more tests around this, and moved the create_all call to happen
after the version check.