Deduplicate attribute key/value strings
We were adding NetCDF file- and variable-level attributes
unconditionally to the ncattributes table. Because of the vast amount
of duplication in these attribute names and values, our only hope to
not blow up the database size was that sqlite itself would adequately
compress things, but this didn't seem to be the case.

We add a new table, to intern the (unique) string values. This needs a
little bit of sqlalchemy machinery to let us seamlessly add attributes
through the usual mapped collection interface. This involves a similar
mechanism to the exsting UniqueMixin, but which is automatically fired
through sqlalchemy's event system.