Added unique attributes which are required to be unique in queries returned from the Database. When the default attribute value is present this is filtered for to provide "least surprise" queries.
Warns when default unique attribute value is not present, but that
attribute is not unique.

Changed attrs argument to not be a mutable dict. Causes errors when
the function is called multiple times.

Removed warning for inconsistent file names in query. With current
diagnostic file naming convention this is no longer correct and will
warn where there is no error.  Removed associated test condition.

Altered data file to contain inconsistent attributes for testing.