Require specified attributes to be unique to ensure disambiguation and consistent data sets returned (#255)
Added argument to getvar and _ncfiles_for_variable to specify
attributes that are required to be unique to ensure data returned
is consistent and correct.

When the default attribute value is present this is filtered for to provide 
"least surprise" queries.

Warns when default unique attribute value is not present, but that
attribute is not unique.

Changed attrs argument to not be a mutable dict. Causes errors when
the function is called multiple times.

Removed warning for inconsistent file names in query. With current
diagnostic file naming convention this is no longer correct and will
warn where there is no error.  Removed associated test condition.

Altered data file to contain inconsistent attributes for testing.

Fixed deprecated pandas.util.testing imports