Added hybrid properties model and is_restart to to NCFile. model infers the realm of the model from the ncfile path. is_restart is a boolean to identify restart variables and also inferred from the ncfile path.
Added hybrid property is_coordinate to CFVariable. This returns a boolean
value which is True if this is a coordinate, which is inferred from the
value of the units.

All hybrid properties are accessible as a python method from a returned
object and as a column in a query.

This is moving logic from explorer into the classes themselves to simplify
and speed up using the explorer GUI elements.

Added inferred properties to get_variables when argument inferred is True.

get_variables can now return *all* variables from *all* experiments if
experiment is None. In this case it will also return the experiment name.

Updated tests, but far from comprehensive.