Reorder attribute cache lookups
The cache lookup was moved to be after a query, because the proper
ordering was breaking not-null constraints. It seems that this was due
to sqlalchemy's propensity to autoflush the session when queries are
issued. This is probably a reasonable optimisation in most cases, but
we rely on being able to normalise the objects within the session
before they're committed to the database.

To allow us to reorder the lookup, we disable autoflush on the
session. This will have no impact on database readers (i.e. most
users). Writers who wish to query the database within the same session
as writing may have to issue a manual flush. Two flushes are added to
the test suite, where this was the case.