Make cell_methods-derived bounds optional in getvar
This is particularly useful for ice data where we can get the time
bounds for averaging intervals. However, because we were attaching the
bounds as an xarray Dataset, we could't serialise a variable straight
back to disk, which was catching people out. Instead, we can make the
bounds optional, only if you want them. In this case, return a full
xarray Dataset, rather than just a DataArray.

Closes #284.